first commit
A C4 Offline Documentation App (incomplete)
So far is just a test and includes only image filters.
Getting a weird error about memory leakage from objects (or views?)
being deleted before all the key values are released (or something like
that?)
Anyways, it works, but it crashes if you try to go through all the
examples (probably due to that memory leakage error).